[measureimagequality]: (3) Move funcitons from FE to module
